+++ tests/Model/Book.php
@@ -17,11 +17,23 @@ class Book
 {
     public function __construct(
         private string $title,
+        private int $numOfPages = 100,
+        private string $author = 'Anonymous',
     ) {
     }
 
     public function getTitle(): string
     {
         return $this->title;
+    }
+
+    public function getNumOfPages(): int
+    {
+        return $this->numOfPages;
+    }
+
+    public function getAuthor(): string
+    {
+        return $this->author;
     }
 }